api/requirements.txt
adapter-transformers==2.2.0
fastapi
python-dotenv
colabcode
pydantic